feat: apply scroll hook to GroupChannel MessageList
