+++ src/smart-components/Channel/components/Message/index.tsx
@@ -31,7 +31,7 @@ type MessageUIProps = {
   hasSeparator?: boolean;
   chainTop?: boolean;
   chainBottom?: boolean;
-  handleScroll: () => void;
+  handleScroll: (event: React.UIEvent<HTMLElement>) => void;
   // for extending
   renderMessage?: (props: RenderMessageProps) => React.ReactElement;
   renderCustomSeparator?: (props: RenderCustomSeparatorProps) => React.ReactElement;
